Rearrange the code so circular imports do not happen
- For db/settings.py though it can be splinto db/settings/structure.py
and db/settings/cache.py the easiest way is to just reduce the surface
of imports that import whole files by moving the constants to their
own dedicated files as it should always be with organized code.

- The long standing get_current_version() + Github() circular import
hit again here, so decided to fix it for good, by rearranging the way
the function works in a way that Github is passed to it only in the 1
place that we want to check for updates. All other places dont need it
so importing it is also not needed.
