+++ rotkehlchen/data_migrations/constants.py
@@ -1,0 +1,1 @@
+LAST_DATA_MIGRATION = 11
