+++ rotkehlchen/utils/misc.py
@@ -382,7 +382,7 @@ def is_production() -> bool:
     if getattr(sys, 'frozen', False) is False:
         return False
 
-    version = get_current_version(check_for_updates=False).our_version
+    version = get_current_version().our_version
     if 'dev' in version:
         return False
